Revert "Revert "添加表面交互和材质基本关联""
This reverts commit 9a5804e03ed177db478ad71bbcb1dbbb60aee7a3.

+++ HyperionFrame/Include/Scenes/Shape.cpp
@@ -29,7 +29,7 @@ AABB Shape::GenerateAABB()
 
 UINT Shape::GetFaceCount()
 {
-	return m_indices.size() / 3;
+	return (UINT)m_indices.size() / 3;
 }
 
 HTriangle Shape::GetFace(UINT faceIndex)
